HW_5 fix v.1

+++ site-form-works/credit_calc/app/forms.py
@@ -3,7 +3,7 @@ from django import forms
 
 class CalcForm(forms.Form):
     initial_fee = forms.IntegerField(label="Стоимость товара")
-    rate = forms.IntegerField(label="Процентная ставка")
+    rate = forms.FloatField(label="Процентная ставка")
     months_count = forms.IntegerField(label="Срок кредита в месяцах")
 
     def clean_initial_fee(self):
@@ -15,14 +15,13 @@ class CalcForm(forms.Form):
 
     def clean_rate(self):
         rate = self.cleaned_data.get('rate')
-        if not rate  or rate  < 0:
+        if not rate or rate < 0:
             raise forms.ValidationError("Процентная ставка не может быть отрицательной")
         return rate
 
-
     def clean_months_count(self):
         months_count = self.cleaned_data.get('months_count')
-        if not months_count or months_count  < 0:
+        if not months_count or months_count < 0:
             raise forms.ValidationError("Необходимо указать количество месяцев")
         return months_count
